# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AGC Cheswerta CMS API",
    "version": "1.0.0",
    "description": "Church management system API \u2014 members, givings, attendance, fellowship scheduling, council roster, SMS broadcast, and admin tools.\n\n**Getting started:** call `POST /api/auth/login`, copy the `token` from the response, then click **Authorize** above and paste it in as `<token>` (no need to type \"Bearer \" \u2014 Swagger adds that automatically). Every other request will then include it.\n\n**Not included:** a few blueprints registered in the app (announcements, modules, feature flags, generic settings, audit log, account locks, login attempts, file uploads, generic db query, stats/PDF reports, SMS logs, church info) aren't documented here yet \u2014 ask to have them added and share those route files.\n\n**Server selector:** the dropdown at the top switches between Local and Production \u2014 edit the Production URL in the spec to match your real domain."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev server (default Flask port)"
    },
    {
      "url": "https://REPLACE-WITH-YOUR-PRODUCTION-DOMAIN",
      "description": "Production \u2014 replace with your real API domain"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from POST /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "You do not have permission to perform this action."
          }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Giving": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "Tithe",
              "Offering",
              "Mission",
              "Baby Center",
              "Special Giving",
              "Other"
            ]
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": 5000
          },
          "member_name": {
            "type": "string",
            "nullable": true,
            "description": "Only used when type is Tithe or Other"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "recorded_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GivingInput": {
        "type": "object",
        "required": [
          "type",
          "amount"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Tithe",
              "Offering",
              "Mission",
              "Baby Center",
              "Special Giving",
              "Other"
            ]
          },
          "amount": {
            "type": "number",
            "format": "float",
            "minimum": 0.01
          },
          "member_name": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Also accepted as 'description'"
          }
        }
      },
...
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Choir"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "leader_name": {
            "type": "string",
            "nullable": true,
            "example": "Mary Wanjiku"
          },
          "member_count": {
            "type": "integer",
            "description": "Live count of Members with this department"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DepartmentInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Choir"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "leader_name": {
            "type": "string",
            "nullable": true,
            "description": "Also accepted as 'leader'"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/givings": {
      "get": {
        "tags": [
          "Givings"
        ],
        "summary": "List givings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Givings"
        ],
        "summary": "Record a giving",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GivingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Giving"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/givings/{giving_id}": {
      "put": {
        "tags": [
          "Givings"
        ],
        "summary": "Update a giving",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "giving_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GivingInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Giving"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Givings"
        ],
        "summary": "Delete a giving",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "giving_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/broadcast/departments": {
      "get": {
        "tags": [
          "Broadcast"
        ],
        "summary": "Distinct department names for the audience picker",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Departments"
          }
        }
      }
    },
...
    "/api/departments": {
      "get": {
        "tags": [
          "Departments"
        ],
        "summary": "List departments with live member counts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Departments"
        ],
        "summary": "Create a department",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/departments/{dept_id}": {
      "put": {
        "tags": [
          "Departments"
        ],
        "summary": "Update a department",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Departments"
        ],
        "summary": "Delete a department (admin/super_admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}